+++ src/hardis-status-provider.ts
@@ -1,18 +1,9 @@
 import * as vscode from "vscode";
-import * as npmApi from "npm-api";
 import * as path from "path";
 import * as GitUrlParse from "git-url-parse";
 import moment = require("moment");
-import {
-  execCommand,
-  execSfdxJson,
-  RECOMMENDED_SFDX_CLI_VERSION,
-} from "./utils";
+import { execSfdxJson } from "./utils";
 import { Logger } from "./logger";
-const npm = new npmApi();
-
-let nodeInstallOk = false;
-let gitInstallOk = false;
 
 export class HardisStatusProvider
   implements vscode.TreeDataProvider<StatusTreeItem>
@@ -52,8 +43,6 @@ export class HardisStatusProvider
         ? await this.getOrgItems({ devHub: true })
         : topic.id === "status-git"
         ? await this.getGitItems()
-        : topic.id === "status-plugins"
-        ? await this.getPluginsItems()
         : [];
     console.timeEnd("TreeViewItem_init_" + topic.id);
     Logger.log("Completed TreeViewItem_init_" + topic.id);
@@ -302,193 +291,6 @@ export class HardisStatusProvider
     return items;
   }
 
-  private async getPluginsItems(): Promise<any[]> {
-    const items: any = [];
-
-    // Check node.js version
-    if (nodeInstallOk === false) {
-      const nodeVersionStdOut: string =
-        (
-          await execCommand("node --version", this, {
-            output: true,
-            fail: false,
-          })
-        ).stdout ||
-        process.env.NODE_PATH ||
-        "error";
-      const nodeVersionMatch = /v([0-9]+)\./gm.exec(nodeVersionStdOut);
-      if (!nodeVersionMatch) {
-        vscode.window
-          .showWarningMessage(
-            "You need Node.js installed on your computer. Please download and install it (version 14 minimum), then restart VsCode",
-            "Download and install Node.js LTS"
-          )
-          .then((selection) => {
-            if (selection === "Download and install Node.js LTS") {
-              vscode.env.openExternal(
-                vscode.Uri.parse("https://nodejs.org/en/")
-              );
-            }
-          });
-      } else if (parseInt(nodeVersionMatch[1]) < 14.0) {
-        vscode.window
-          .showWarningMessage(
-            `You have a too old version (${nodeVersionMatch[1]}) of Node.js installed on your computer. Please download and install it (version 14 minimum), then restart VsCode`,
-            "Download and install Node.js LTS"
-          )
-          .then((selection) => {
-            if (selection === "Download and install Node.js LTS") {
-              vscode.env.openExternal(
-                vscode.Uri.parse("https://nodejs.org/en/")
-              );
-            }
-          });
-      } else {
-        nodeInstallOk = true;
-      }
-    }
-
-    // Check git version
-    if (gitInstallOk === false) {
-      const gitVersionStdOut: string =
-        (
-          await execCommand("git --version", this, {
-            output: true,
-            fail: false,
-          })
-        ).stdout || "error";
-      const gitVersionMatch = /git version ([0-9]+)\./gm.exec(gitVersionStdOut);
-      if (!gitVersionMatch) {
-        vscode.window
-          .showWarningMessage(
-            "You need Git installed on your computer. Please download and install it (select GIT BASH in options), then restart VsCode",
-            "Download and install Git"
-          )
-          .then((selection) => {
-            if (selection === "Download and install Git") {
-              vscode.env.openExternal(
-                vscode.Uri.parse("https://git-scm.com/downloads")
-              );
-            }
-          });
-      } else {
-        gitInstallOk = true;
-      }
-    }
-
-    // Check sfdx related installs
-    const plugins = [
-      { name: "sfdx-hardis" },
-      { name: "sfdx-essentials" },
-      { name: "sfpowerkit" },
-      { name: "sfdmu" },
-      { name: "sfdx-git-delta" },
-      { name: "texei-sfdx-plugin" },
-    ];
-    const outdated: any[] = [];
-    // check sfdx-cli version
-    const sfdxCliVersionStdOut: string = (
-      await execCommand("sfdx --version", this, { output: true, fail: false })
-    ).stdout;
-    const sfdxCliVersionMatch = /sfdx-cli\/([^\s]+)/gm.exec(
-      sfdxCliVersionStdOut
-    );
-    let sfdxCliVersion = "(missing)";
-    if (sfdxCliVersionMatch) {
-      sfdxCliVersion = sfdxCliVersionMatch[1];
-    }
-    const latestSfdxCliVersion = await npm.repo("sfdx-cli").prop("version");
-    const config = vscode.workspace.getConfiguration("vsCodeSfdxHardis");
-    const recommendedSfdxCliVersion =
-      config.get("ignoreSfdxCliRecommendedVersion") === true
-        ? latestSfdxCliVersion
-        : RECOMMENDED_SFDX_CLI_VERSION || latestSfdxCliVersion;
-    const sfdxCliItem = {
-      id: `sfdx-cli-info`,
-      label: `sfdx-cli v${sfdxCliVersion}`,
-      command: "",
-      tooltip: `Recommended version of sfdx-cli is installed`,
-      icon: "success.svg",
-    };
-    let sfdxCliOutdated = false;
-    if (sfdxCliVersion !== recommendedSfdxCliVersion) {
-      sfdxCliOutdated = true;
-      sfdxCliItem.label =
-        sfdxCliItem.label.includes("missing") &&
-        !sfdxCliItem.label.includes("(link)")
-          ? sfdxCliItem.label
-          : sfdxCliItem.label + " (upgrade available)";
-      sfdxCliItem.command = `npm install sfdx-cli@${recommendedSfdxCliVersion} -g`;
-      sfdxCliItem.tooltip = `Click to upgrade sfdx-cli to ${recommendedSfdxCliVersion}`;
-      sfdxCliItem.icon = "warning.svg";
-    }
-    items.push(sfdxCliItem);
-    // get currently installed plugins
-    const sfdxPlugins =
-      (await execCommand("sfdx plugins", this, { output: true, fail: false }))
-        .stdout || "";
-    // Check installed plugins status version
-    const pluginPromises = plugins.map(async (plugin) => {
-      // Check latest plugin version
-      const latestPluginVersion = await npm.repo(plugin.name).prop("version");
-      let pluginLabel = plugin.name;
-      const regexVersion = new RegExp(`${plugin.name} (.*)`, "gm");
-      const versionMatches = [...sfdxPlugins.matchAll(regexVersion)];
-      if (versionMatches.length > 0) {
-        pluginLabel += ` v${versionMatches[0][1]}`;
-      } else {
-        pluginLabel += " (missing)";
-      }
-      const pluginItem = {
-        id: `plugin-info-${plugin.name}`,
-        label: pluginLabel,
-        command: "",
-        tooltip: `Latest version of SFDX plugin ${plugin.name} is installed`,
-        icon: "success.svg",
-      };
-      if (!sfdxPlugins.includes(`${plugin.name} ${latestPluginVersion}`)) {
-        pluginItem.label =
-          pluginItem.label.includes("missing") &&
-          !pluginItem.label.includes("(link)")
-            ? pluginItem.label.replace("(link)", "(localdev)")
-            : pluginItem.label + " (upgrade available)";
-        pluginItem.command = `echo y|sfdx plugins:install ${plugin.name}`;
-        pluginItem.tooltip = `Click to upgrade SFDX plugin ${plugin.name} to ${latestPluginVersion}`;
-        if (!pluginItem.label.includes("(localdev)")) {
-          pluginItem.icon = "warning.svg";
-          outdated.push(plugin);
-        }
-      }
-      items.push(pluginItem);
-    });
-    // Await parallel promises to be completed
-    await Promise.all(pluginPromises);
-    // Propose user to upgrade if necessary
-    if (outdated.length > 0) {
-      vscode.window
-        .showWarningMessage(
-          "Some plugins are not up to date, please click to upgrade, then wait for the process to be completed before performing actions",
-          "Upgrade plugins"
-        )
-        .then((selection) => {
-          if (selection === "Upgrade plugins") {
-            let command = outdated
-              .map((plugin) => `echo y|sfdx plugins:install ${plugin.name}`)
-              .join(" && ");
-            if (sfdxCliOutdated === true) {
-              command = "npm install sfdx-cli -g && " + command;
-            }
-            command = command + ` && sfdx hardis:work:ws --event refreshStatus`;
-            vscode.commands.executeCommand(
-              "vscode-sfdx-hardis.execute-command",
-              command
-            );
-          }
-        });
-    }
-    return items.sort((a: any, b: any) => (a.label > b.label ? 1 : -1));
-  }
-
   /**
    * List all topics
    */
@@ -544,12 +346,6 @@ export class HardisStatusProvider
         id: "status-git",
         label: "Git Status",
         icon: "git.svg",
-        defaultExpand: true,
-      },
-      {
-        id: "status-plugins",
-        label: "Plugins",
-        icon: "plugins.svg",
         defaultExpand: true,
       },
       {